requirements.txt
numpy
transformers
sentencepiece
Pillow
ipykernel